Added profile class
