+++ .gitignore
@@ -2,3 +2,4 @@
 *.pyc
 *.png
 *.jpg
+*.idea
